Update tests after moving ContestCategory out of StringUtil.
